GearTrade/requirements.txt
Django==5.2
djangorestframework==3.16.0
pillow==11.2.1
django-filter==25.1
python-dotenv==1.1.0
openai==1.82.0
